Verify token, Role-based authorization, Generate random user password

+++ backend/Controllers/HospitalController.cs
@@ -12,7 +12,7 @@ namespace backend.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
-    [Authorize]
+    [Authorize(Roles = "admin")]
     public class HospitalController : ControllerBase
     {
         private readonly IHospitalRepository _hospitalRepository;
@@ -21,7 +21,6 @@ namespace backend.Controllers
         {
             _hospitalRepository = hospitalRepository;
         }
-        
 
         [HttpPost]
         public async Task<IActionResult> Create(Hospital hospital)
@@ -77,6 +76,12 @@ namespace backend.Controllers
             if (!ObjectId.TryParse(id, out _)) return NotFound("Invalid ID");
             try
             {
+                var exist = await _hospitalRepository.Get(id);
+                if (exist == null)
+                {
+                    throw new Exception();
+                }
+
                 var result = await _hospitalRepository.Update(id, hospital);
                 return Ok(result);
             }
@@ -93,6 +98,12 @@ namespace backend.Controllers
             if (!ObjectId.TryParse(id, out _)) return NotFound("Invalid ID");
             try
             {
+                var exist = await _hospitalRepository.Get(id);
+                if (exist == null)
+                {
+                    throw new Exception();
+                }
+
                 var result = await _hospitalRepository.Delete(id);
                 return Ok(result);
             }
